Mute some of the errors "Implicit conversion from type.."

+++ core/object/class_db.cpp
@@ -427,8 +427,8 @@ uint64_t ClassDB::get_api_hash(APIType p_api) {
 			snames.sort_custom<StringName::AlphCompare>();
 
 			for (List<StringName>::Element *F = snames.front(); F; F = F->next()) {
-				hash = hash_djb2_one_64(F->get().hash(), hash);
-				hash = hash_djb2_one_64(t->constant_map[F->get()], hash);
+				hash = hash_djb2_one_64(uint64_t(F->get().hash()), hash);
+				hash = hash_djb2_one_64(uint64_t(t->constant_map[F->get()]), hash);
 			}
 		}
